requirements.txt
adal==1.2.7
ansible==9.13.0
ansible-core==2.16.14
antlr4-python3-runtime==4.13.2
appdirs==1.4.4
applicationinsights==0.11.9
argcomplete==3.5.3
azure-appconfiguration==1.1.1
azure-batch==14.2.0
azure-cli==2.63.0
azure-cli-core==2.63.0
azure-cli-telemetry==1.1.0
azure-common==1.1.28
azure-core==1.30.2
azure-cosmos==3.2.0
azure-data-tables==12.4.0
azure-datalake-store==0.0.53
azure-graphrbac==0.60.0
azure-keyvault-administration==4.4.0b2
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.9.0b3
azure-keyvault-secrets==4.7.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-appconfiguration==3.0.0
azure-mgmt-appcontainers==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-batch==17.3.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0b3
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-compute==31.0.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-containerservice==31.0.0
azure-mgmt-core==1.3.2
azure-mgmt-cosmosdb==9.5.1
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datamigration==10.0.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==10.2.0b2
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0
azure-mgmt-imagebuilder==1.3.0
azure-mgmt-iotcentral==10.0.0b1
azure-mgmt-iothub==3.0.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==1.0.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==5.0.0
azure-mgmt-msi==7.0.0
azure-mgmt-netapp==10.1.0
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.2.0b17
azure-mgmt-recoveryservices==3.0.0
azure-mgmt-recoveryservicesbackup==9.1.0
azure-mgmt-redhatopenshift==1.4.0
azure-mgmt-redis==14.3.0
azure-mgmt-resource==23.1.1
azure-mgmt-search==9.0.0
azure-mgmt-security==6.0.0
azure-mgmt-servicebus==8.2.0
azure-mgmt-servicefabric==2.1.0
azure-mgmt-servicefabricmanagedclusters==2.0.0b6
azure-mgmt-servicelinker==1.2.0b2
azure-mgmt-signalr==2.0.0b1
azure-mgmt-sql==4.0.0b17
azure-mgmt-sqlvirtualmachine==1.0.0b5
azure-mgmt-storage==21.2.0
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.2.0
azure-monitor-query==1.3.0
azure-multiapi-storage==1.2.0
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.19.0
azure-synapse-managedprivateendpoints==0.4.0
azure-synapse-spark==0.2.0
babel==2.17.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blivet==3.11.0
blivet-gui==2.6.0
boto3==1.36.12
botocore==1.36.12
Brlapi==0.8.5
certifi==2023.5.7
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
colorama==0.4.6
crypt_r==3.13.1
cryptography==43.0.0
cson==0.8
cssselect==1.1.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
decorator==5.1.1
Deprecated==1.2.15
distro==1.9.0
dnf==4.22.0
fabric==3.2.2
fedora-third-party==0.10
file-magic==0.4.0
fluidity-sm==0.2.0
git-filter-repo==2.47.0
humanfriendly==10.0
humanize==4.9.0
idna==3.7
initial-setup==0.3.101
inkex==1.4.0
invoke==2.2.0
iso639==0.1.4
isodate==0.6.1
javaproperties==0.8.1
Jinja2==3.1.5
jmespath==1.0.1
jsondiff==2.2.1
knack==0.12.0
langtable==0.0.68
lexicon==2.0.1
libcomps==0.1.21
libdnf==0.73.4
libdnf5==5.2.8.1
louis==3.28.0
lxml==5.2.1
MarkupSafe==2.1.5
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4
nftables==0.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
packaging==24.2
paramiko==3.5.0
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
pkginfo==1.11.2
ply==3.11
portalocker==2.10.1
productmd==1.43
progressbar2==3.53.2
psutil==5.9.8
ptyprocess==0.7.0
pwquality==1.4.5
pyasn1==0.6.0
pycairo==1.25.1
pycomposefile==0.0.30
pycparser==2.20
pycups==2.0.4
pyenchant==3.2.2
pygdbmi==0.11.0.0
pygit2==1.17.0
PyGithub==2.3.0
Pygments==2.18.0
PyGObject==3.48.2
pyinotify==0.9.6
PyJWT==2.8.0
pykickstart==3.58
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyparted==3.13.0
PyQt5==5.15.10
PyQt5_sip==12.15.0
pyserial==3.5
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-linux-procfs==0.7.3
python-meh==0.52
python-utils==3.7.0
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
regex==2024.9.11
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
requests-oauthlib==1.3.1
resolvelib==1.0.1
rpm==4.20.0
rpmautospec==0.7.3
rpmautospec-core==0.1.5
s3transfer==0.11.2
scour==0.38.2
scp==0.14.5
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
semver==3.0.2
sentry-sdk==2.17.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.8.2
soupsieve==2.6
speg==0.3
sshtunnel==0.4.0
systemd-python==235
tabulate==0.9.0
tinycss2==1.3.0
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
wrapt==1.17.1
xkbregistry==0.3
xmltodict==0.13.0
